CLI-575: Fix stalled tests
